build: Configure workflow Unix

+++ upnp/src/threadutil/ThreadPool.cpp
@@ -426,8 +426,8 @@ static void SetSeed(void)
         srand((unsigned int)t.tv_usec +
                 (unsigned int)ithread_get_current_thread_id().p);
 #elif defined(BSD) || defined(__APPLE__) || defined(__FreeBSD_kernel__)
-        srand(t.tv_usec +
-                ithread_get_current_thread_id());
+        srand((unsigned long)t.tv_usec +
+                (unsigned long)ithread_get_current_thread_id());
 #elif defined(__linux__) || defined(__sun) || defined(__CYGWIN__) || \
         defined(__GLIBC__)
         srand((unsigned int)t.tv_usec +
